Move probes to their own module
